Vola-Modell/Modelle/Option_Pricer/hydra-pricer/requirements.txt
numpy==1.26.4
scipy==1.11.4
pandas==2.1.3
plotly==5.18.0
streamlit==1.29.0
yfinance==0.2.33
pytest==7.4.3
black==23.11.0
fpdf2==2.7.9
